Split the NoiseModelFactorN to individual header
