+++ gtsam/navigation/ConstantVelocityFactor.h
@@ -19,6 +19,7 @@
 
 #include <gtsam/navigation/NavState.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 namespace gtsam {
