+++ gtsam/nonlinear/ExtendedPriorFactor.h
@@ -17,6 +17,7 @@
 
 #include <gtsam/linear/NoiseModel.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 #include <optional>
