+++ gtsam/nonlinear/FunctorizedFactor.h
@@ -19,6 +19,7 @@
 
 #include <gtsam/base/Testable.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 #include <cmath>
