+++ gtsam/sfm/TranslationFactor.h
@@ -23,6 +23,7 @@
 #include <gtsam/geometry/Unit3.h>
 #include <gtsam/linear/NoiseModel.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 namespace gtsam {
