+++ gtsam/slam/EssentialMatrixConstraint.h
@@ -19,6 +19,7 @@
 #pragma once
 
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <gtsam/geometry/EssentialMatrix.h>
 
 namespace gtsam {
