+++ gtsam/slam/PoseTranslationPrior.h
@@ -11,6 +11,7 @@
 
 #include <gtsam/geometry/concepts.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 namespace gtsam {
