+++ gtsam/slam/RotateFactor.h
@@ -8,6 +8,7 @@
 #pragma once
 
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <gtsam/geometry/Rot3.h>
 
 namespace gtsam {
