+++ gtsam_unstable/dynamics/Pendulum.h
@@ -10,6 +10,7 @@
 #pragma once
 
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 namespace gtsam {
