+++ gtsam_unstable/dynamics/VelocityConstraint.h
@@ -8,6 +8,7 @@
 
 #include <gtsam/base/numericalDerivative.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <gtsam_unstable/dynamics/PoseRTV.h>
 
 #include <cassert>
