+++ gtsam_unstable/dynamics/VelocityConstraint3.h
@@ -7,6 +7,7 @@
 #pragma once
 
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 namespace gtsam {
