+++ gtsam_unstable/slam/GaussMarkov1stOrderFactor.h
@@ -17,6 +17,7 @@
 #pragma once
 
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <gtsam/linear/GaussianFactor.h>
 #include <gtsam/linear/NoiseModel.h>
 #include <gtsam/base/Testable.h>
