+++ gtsam_unstable/slam/LocalOrientedPlane3Factor.h
@@ -9,6 +9,7 @@
 
 #include <gtsam/geometry/OrientedPlane3.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <gtsam_unstable/dllexport.h>
 
 #include <string>
