+++ gtsam_unstable/slam/PartialPriorFactor.h
@@ -18,6 +18,7 @@
 #pragma once
 
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <gtsam/base/Lie.h>
 
 #include <cassert>
