+++ gtsam_unstable/slam/PoseBetweenFactor.h
@@ -18,6 +18,7 @@
 #include <ostream>
 
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <gtsam/geometry/concepts.h>
 #include <gtsam/base/Testable.h>
