+++ gtsam_unstable/slam/PoseToPointFactor.h
@@ -12,6 +12,7 @@
 #include <gtsam/geometry/Point3.h>
 #include <gtsam/geometry/Pose3.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 #include <ostream>
 
 namespace gtsam {
