+++ gtsam_unstable/slam/RelativeElevationFactor.h
@@ -12,6 +12,7 @@
 #include <gtsam_unstable/dllexport.h>
 #include <gtsam/geometry/Pose3.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 namespace gtsam {
