+++ gtsam_unstable/slam/TSAMFactors.h
@@ -20,6 +20,7 @@
 
 #include <gtsam/geometry/Pose2.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
 
 namespace gtsam {
