+++ tests/simulated3D.h
@@ -23,7 +23,8 @@
 #include <gtsam/geometry/Point3.h>
 #include <gtsam/linear/VectorValues.h>
 #include <gtsam/nonlinear/NonlinearFactor.h>
-#include "gtsam/base/OptionalJacobian.h"
+#include <gtsam/nonlinear/NoiseModelFactorN.h>
+#include <gtsam/base/OptionalJacobian.h>
 
 // \namespace
